// src/apps/users/enums/user-error-keys.enum.ts
export enum UserMessagesEnum {
  NOT_FOUND = 'users.not_found.error',
  CREATE_USER_ERROR = 'users.create_user.error',
  ALREADY_EXIST = 'users.already_exist.error',
  EMAIL_INVALID = 'users.email_invalid.error',
  PASSWORD_INVALID = 'users.password_invalid.error',
  PASSWORD_INVALID_LENGHT = 'users.password_invalid_length.error',
  IMAGE_PROFILE_INVALID = 'users.image_profile_invalid.error',
  USER_NAME_INVALID = 'users.user_name_invalid.error',
  PHONE_INVALID = 'users.phone_invalid.error',
  ENABLED_INVALID = 'users.enabled_invalid.error',
  KEY_REFRESH_TOKEN_INVALID = 'users.language_invalid.error',
  INTERNAL_INVALID = 'users.internal_invalid.error',
  PERMISSIONS_UUIDS_ARE_INVALID = 'users.permissions_uuids_invalid.error',
  ROLES_UUIDS_ARE_INVALID = 'users.roles_uuids_invalid.error',
  USER_UUID_IS_INVALID = 'users.user_uuid_invalid.error',
  USER_PHONE_IS_INVALID = 'users.user_phone_invalid.error',
  RECOVERY_PASSWORD_CODE_INVALID = 'users.recovery_passwor_invalid.error',
  RECOVERY_PASSWORD_METHOD_INVALID = 'users.recovery_passwor_method_invalid.error',
  RECOVERY_PASSWORD_CODE_EXPIRED = 'users.recovery_passwor_expired.error',
  RECOVERY_PASSWORD_TOKEN_INVALID = 'users.recovery_password_token_invalid.error',
  RECOVERY_PASSWORD_TOKEN_EXPIRED = 'users.recovery_password_token_expired.error',
  RECOVERY_PASSWORD_NOT_AVAILABLE = 'users.recovery_password_not_available.error',
  REGISTER_ME_NOT_AVAILABLE = 'users.register_me_not_available.error',
  OBJECT_IS_INVALID = 'users.object_is_invalid.error',
  DEVICE_TOKEN_INVALID = 'users.device_token_invalid.error',
  DEVICE_UUID_INVALID = 'users.device_uuid_invalid.error',
  PERMISSIONS_INVALID = 'users.permissions_invalid.error',
  ROLES_INVALID = 'users.roles_invalid.error',
  ACTIVE_INVALID = 'users.active_invalid.error',
}